# [go/mysql] Avoid static buffer and use tiered pool in 