Benchmark results:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkParallelShortQueries-8      3259          3368          +3.34%
BenchmarkParallelMediumQueries-8     6444          6561          +1.82%
BenchmarkParallelRandomQueries-8     5963212       5441551       -8.75%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParallelShortQueries-8      3.37         3.27         0.97x
BenchmarkParallelMediumQueries-8     2543.26      2497.78      0.98x

benchmark                            old allocs     new allocs     delta
BenchmarkParallelShortQueries-8      23             23             +0.00%
BenchmarkParallelMediumQueries-8     9              8              -11.11%
BenchmarkParallelRandomQueries-8     14             14             +0.00%

benchmark                            old bytes     new bytes     delta
BenchmarkParallelShortQueries-8      720           812           +12.78%
BenchmarkParallelMediumQueries-8     27205         22720         -16.49%
BenchmarkParallelRandomQueries-8     35898388      29652823      -17.40%
```

I added tiered pool here because without it performance hit is too big (apparently because we use smaller buffer for readOnePacket or something and then when we read query with very high probability buffer from the pool is that smaller buffer).
cc @danieltahara