@sougou 2018-09-05
I'll clarify my comments in the other PR