@LK4D4 2018-09-05
So, here is benchmark results comparing to tiered pool PR:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkParallelShortQueries-8      3726          3771          +1.21%
BenchmarkParallelMediumQueries-8     6016          7715          +28.24%
BenchmarkParallelRandomQueries-8     5432915       5531186       +1.81%

benchmark                            old MB/s     new MB/s     speedup
BenchmarkParallelShortQueries-8      2.95         2.92         0.99x
BenchmarkParallelMediumQueries-8     2724.08      2124.35      0.78x

benchmark                            old allocs     new allocs     delta
BenchmarkParallelShortQueries-8      23             23             +0.00%
BenchmarkParallelMediumQueries-8     8              8              +0.00%
BenchmarkParallelRandomQueries-8     14             14             +0.00%

benchmark                            old bytes     new bytes     delta
BenchmarkParallelShortQueries-8      720           815           +13.19%
BenchmarkParallelMediumQueries-8     20886         23261         +11.37%
BenchmarkParallelRandomQueries-8     29717958      30225503      +1.71%
```

And I don't know what could cause this slowdown... Maybe I'll try to revert to readPacketDirect...