@sougou 2018-09-06
What are the benchmarks compared against?