@sougou 2018-09-06
I think the reason why medium buffer is slower is because the size is just above the maxSize. So, it ends up with fresh allocations for every iteration.

I spent some more time digging through the code, and have a few observations. The TL;DR is that we can get rid of all read and write policies.

* For read policy, MaxPacketSize is 16M. So, big buffers will never be put back into the pool because size exceeds maxSize. This means that we don't need to track policies.
* We need more buckets. I think we should go up to a maxSize of 15M. Beware of benchmarks, because a random number will skew it towards the huge maxSize. You may have to exponentiate the random numbers, or use different bands.
* For writes, there is a readability problem: the different code paths are not due to policy. They are actually due to the packet size. So, the code should be changed to explicitly reflect that. This means that we can get rid of write policies.
* bufio.Writer has a `Reset` function which allows you to change the writer. This means that you can put it in a sync.Pool.
* Also, the use of `direct` is bug prone. It should not be a flag passed into  `writeEphemeralPacket`. Instead, there should be a separate set of functions that cleanly encapsulate the use of bufio with a defer that flushes it (and recycles if we use sync.Pool). This should probably be a separate PR.