@sougou 2018-09-06
I got confused (by looking at the Pool benchmarks). I see that you're now using 16MB.

The buffer thing is weird indeed. We can look at varying the size of the medium query size to see where it jumps. That could give us a clue.

I'm wondering if pprof would help for such short runs.