@LK4D4 2018-09-06
@sougou I tried twice bigger queries with the same result :(
Btw, adding `_ [connBufferSize]byte` to struct "fixes" the problem. Apparently some gc tricks again.